streamlit_app/requirements.txt
streamlit
pandas
requests
google-cloud-aiplatform
matplotlib
seaborn
